Add User Coupon Searching
Added the functionality for users to ONLY see coupons near them. Also attached the function that they will only see the redacted version that does not include any features allowing for redemption to occur. Also created function to allow override to get a list of coupons at a specified location.