Changed viewing criteria
As per our conversation in class, added a  centerLat and centerLong field to the coupons so we know the center location. Also changed the user doc to have a viewWindow object that should have the same lat/long combo of 4 fields as the coupon doc. Modified the publish function to handle this new data as well.