Failed attempt
I have written the code to handle redemption, however I am having a lot of errors testing it. It tells me that something does not match the expected value or type on a user find. By commenting out everything I have narrowed it down to a user database error. It may extend to coupons but I have been unable to test that portion yet...