Fixed bugs
Fixed all the of Async errors and changed the way the finds worked so the database could find the correct information successfully. Added a decent amount of test code for collecting and redeeming to ensure that they work. Still need to stress-test to ensure behavior persists in large throughput. Redemption return needs to be tested on the client-side still but gives undefined on server side due to timeout.